refactor: extract SegmentStrippingResolver to a shared spring-utils class
The behavioral fix added the resolver as a nested class duplicated in
WebappWebMvcConfig and PhysicalTenantWebMvcConfig. Promote it to a single
public class in spring-utils, next to PhysicalTenantContext — both exist to
handle the physical-tenant path prefix, so that is its conceptual home, and
both consuming modules already depend on spring-utils.

spring-utils already declares its servlet/web APIs at provided scope so it
can host request-scoped web helpers without forcing a web stack onto non-web
consumers; spring-webmvc (for PathResourceResolver) is added the same way.
The PT-specific reasoning stays as a comment at each call site.
